refactor: move normalize_service_date to normalize.py (#399)
normalize_service_date (and its _DATE_PARTS_RE) lived in pptx_reader, forcing
the data layer (db.py) to import the PPTX-parsing layer. Move it to normalize.py
(a pure string utility). db.py now imports from normalize; pptx_reader re-imports
it (still used internally and by existing tests). Adds a layering guard test.
